# server/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"notes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Expense": {
				"properties": {
					"id": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"vendor": {
						"type": "string",
						"nullable": true
					},
					"category_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"date",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateExpenseRequest": {
				"properties": {
					"date": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"vendor": {
						"type": "string"
					},
					"category_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"amount",
					"vendor",
					"category_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReportRow": {
				"properties": {
					"periodStart": {
						"type": "string"
					},
					"groupKey": {
						"type": "string"
					},
					"groupLabel": {
						"type": "string"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"expenseCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"periodStart",
					"groupKey",
					"groupLabel",
					"totalAmount",
					"expenseCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Report": {
				"properties": {
					"aggregation": {
						"type": "string",
						"enum": [
							"daily",
							"weekly",
							"monthly",
							"yearly"
						]
					},
					"groupBy": {
						"type": "string",
						"enum": [
							"category",
							"vendor"
						]
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/ReportRow"
						},
						"type": "array"
					}
				},
				"required": [
					"aggregation",
					"groupBy",
					"from",
					"to",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestInput": {
				"properties": {
					"aggregation": {
						"type": "string",
						"enum": [
							"daily",
							"weekly",
							"monthly",
							"yearly"
						]
					},
					"groupBy": {
						"type": "string",
						"enum": [
							"category",
							"vendor"
						]
					},
					"filterBy": {
						"properties": {
							"categoryIds": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						},
						"type": "object"
					},
					"from": {
						"type": "string"
					},
					"to": {
						"type": "string"
					}
				},
				"required": [
					"aggregation",
					"groupBy",
					"from",
					"to"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "expense-tracker-server",
		"version": "0.1.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/expenses": {
			"get": {
				"operationId": "GetExpenses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Expense"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortDirection",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "vendor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateExpense",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Expense"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExpenseRequest"
							}
						}
					}
				}
			}
		},
		"/api/expenses/{id}": {
			"get": {
				"operationId": "GetExpense",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Expense"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateExpense",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Expense"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExpenseRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExpense",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/reporting": {
			"post": {
				"operationId": "GetExpensesReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Report"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestInput"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}